coverage: Optimize code block storage + load_offset support
Now info about all executed code blocks stored in list, which is not
optimized because a lot of code blocks may be executed in loop.
So, instead store info about executed blocks will be stored in the map with
pair of address/size as a key + info will be store extra |count|
parameter to know how many times this block was executed.

Also, use load_offset from Image to modify Module start address to match
an original ELF address.

Change-Id: I68278253550aaeebe81c38b68ea816b7aed8f883