fuzzing: Updates for SysFuzzer
Introduce 'arg_value_checker' argument in the SysFuzzer callback
context.
'arg_value_checker' is a callback to do modification/validation of
an argument parsed value. Example, if a caller knows that valid value
for the argument in in the range 0..100, 'arg_value_checker' can be used
by a fuzzer to narrow a value to this range.

Change-Id: I54a254ca21ebc083d1554624b08b2b7c2a6fc2ca